requirements.txt
spotipy
telethon
python-dotenv